2.0.0-SNAPSHOT Upload-1

+++ api/src/main/java/br/com/munif/framework/vicente/api/BaseAPI.java
@@ -6,31 +6,25 @@
 package br.com.munif.framework.vicente.api;
 
 import br.com.munif.framework.vicente.application.BaseService;
-import br.com.munif.framework.vicente.core.Utils;
+import br.com.munif.framework.vicente.core.ReflectionUtil;
 import br.com.munif.framework.vicente.core.VicQuery;
 import br.com.munif.framework.vicente.core.VicReturn;
 import br.com.munif.framework.vicente.domain.BaseEntity;
-import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
-
-import org.jetbrains.annotations.NotNull;
+import br.com.munif.framework.vicente.domain.BaseEntityHelper;
 import org.springframework.context.annotation.Scope;
-import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
- *
  * @author munif
  */
 @RestController
@@ -44,75 +38,91 @@ public class BaseAPI<T extends BaseEntity> {
     }
 
     @Transactional
-    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public T delete(@PathVariable String id) {
-        T entity = service.view(id);
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity<T> delete(@PathVariable String id) {
+        T entity = service.load(id);
+        if (entity == null) {
+            throw new VicenteNotFoundException("Not found");
+        }
+        if (!entity.canDelete()) {
+            throw new VicenteRightsException("DELETE," + id + "," + entity.r());
+        }
+        beforeDelete(entity);
         service.delete(entity);
-        return entity;
+        return new ResponseEntity<>(entity, HttpStatus.NO_CONTENT);
     }
 
     @Transactional
-    @RequestMapping(method = RequestMethod.POST)
+    @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
-    public T save(@RequestBody @Valid T model) {
+    public ResponseEntity<T> save(@RequestBody @Valid T model) {
         beforeSave(model);
-        if (service.view(model.getId()) != null) {
+        if (service.load(model.getId()) != null) {
             throw new VicenteCreateWithExistingIdException("create With Existing Id=" + model.getId());
         }
 
         T entity = service.save(model);
-        return entity;
+        return new ResponseEntity<>(entity, HttpStatus.CREATED);
     }
 
     @Transactional
-    @RequestMapping(value = "/{id}", method = RequestMethod.PUT, consumes = "application/json")
-    public T update(@PathVariable("id") String id, @RequestBody @Valid T model) {
-        beforeUpdate(id, model);
-        T entity = service.save(model);
-        return entity;
+    @PutMapping(value = "", consumes = "application/json")
+    public ResponseEntity<T> updateWithoutId(@RequestBody @Valid T model) {
+        return doUpdate(model);
+
     }
 
     @Transactional
-    @RequestMapping(value = "", method = RequestMethod.PUT, consumes = "application/json")
-    public ResponseEntity<T> update2(@RequestBody @Valid T model) {
+    @PutMapping(value = "/{id}", consumes = "application/json")
+    public ResponseEntity<T> updateWithId(@PathVariable("id") String id, @RequestBody @Valid T model) {
+        model.setId(id);
+        return doUpdate(model);
+
+    }
+
+    @Transactional
+    @PatchMapping(value = "/{id}", consumes = "application/json")
+    public ResponseEntity<Void> patch(@RequestBody @Valid Map model) {
+        service.patch(model);
+        return ResponseEntity.noContent().build();
+    }
+
+    @Transactional
+    @PatchMapping(value = "/returning/{id}", consumes = "application/json")
+    public ResponseEntity patchReturning(@RequestBody @Valid Map model) {
+        return ResponseEntity.ok(service.patchReturning(model));
+    }
+
+    private ResponseEntity<T> doUpdate(T model) {
         try {
+            T entity = null;
             HttpStatus ht = HttpStatus.OK;
-            if (service.view(model.getId()) != null) {
-                model = update(model.getId(), model);
+            T oldEntity = service.load(model.getId());
+            if (oldEntity != null) {
+                beforeUpdate(model.getId(), model);
+                BaseEntityHelper.overwriteJsonIgnoreFields(model, oldEntity);
+                entity = service.save(model);
             } else {
-                model = service.save(model);
+                beforeSave(model);
+
+                entity = service.save(model);
                 ht = HttpStatus.CREATED;
             }
-            return new ResponseEntity(model, ht);
+            return new ResponseEntity(entity, ht);
         } catch (Exception ex) {
             ex.printStackTrace();
             return null;
         }
-
     }
 
-    protected void beforeSave(T model) {
-
-    }
-
-    protected void beforeUpdate(String id, T model) {
-
-    }
-
-//    @Transactional
-//    @RequestMapping(method = RequestMethod.GET)
-//    public VicReturn<T> findAll() {
-//        List<T> findAll = service.findAll();
-//        return new VicReturn<T>(findAll);
-//    }
     @Transactional
-    @RequestMapping(method = RequestMethod.GET)
+    @GetMapping(produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public VicReturn<T> findHQL(HttpServletRequest request, VicQuery query) {
         return getVicReturnByQuery(query);
     }
 
     @Transactional
-    @RequestMapping(value = "/vquery", method = RequestMethod.POST)
+    @PostMapping(value = "/vquery", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public VicReturn<T> findVQuery(@RequestBody VicQuery query) {
         return getVicReturnByQuery(query);
     }
@@ -133,21 +143,48 @@ public class BaseAPI<T extends BaseEntity> {
         if (hasMore) {
             result.remove(maxResults);
         }
+        if (query.getQuery() != null && query.getQuery().getFields() != null) {
+            String[] fields = query.getQuery().getFields();
+            List<Map<String, Object>> collect = result.stream().map(s -> getFields(fields, s)).collect(Collectors.toList());
+            return new VicReturn(collect, collect.size(), query.getFirstResult(), hasMore);
+        }
         return new VicReturn<T>(result, result.size(), query.getFirstResult(), hasMore);
     }
 
     @Transactional
-    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public T load(@PathVariable String id) {
-        T view = service.view(id);
-        if (view == null) {
+    @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity load(@PathVariable String id, @RequestParam(required = false) String fields) {
+        T view = service.load(id);
+        if (view == null || !view.canRead()) {
             throw new VicenteNotFoundException("Not found");
-        };
+        }
+
         beforeReturnOne(view);
-        return view;
+        if (fields != null) {
+            Map<String, Object> stringObjectMap = getFields(fields, view);
+            return new ResponseEntity(stringObjectMap, HttpStatus.OK);
+        }
+        return ResponseEntity.ok(view);
     }
 
-    @RequestMapping(value = "/new", method = RequestMethod.GET)
+    private Map<String, Object> getFields(String fields, T view) {
+        String[] split = fields.split(",");
+        return getFields(split, view);
+    }
+
+    private Map<String, Object> getFields(String[] fields, T view) {
+        return ReflectionUtil.objectFieldsToMap(fields, view);
+    }
+
+    @ResponseBody
+    @Transactional
+    @GetMapping(value = "/draw/{id}", produces = "image/svg+xml")
+    public ResponseEntity<String> draw(@PathVariable String id) {
+        String svg = service.draw(id);
+        return ResponseEntity.ok(svg);
+    }
+
+    @GetMapping(value = "/new")
     public T initialState() {
         return service.newEntity();
     }
@@ -156,8 +193,16 @@ public class BaseAPI<T extends BaseEntity> {
         return 20;
     }
 
-    protected void beforeReturnOne(T view) {
+    protected void beforeSave(T model) {
+    }
 
+    protected void beforeUpdate(String id, T model) {
+    }
+
+    protected void beforeReturnOne(T view) {
+    }
+
+    protected void beforeDelete(T entity) {
     }
 
 }
